fix: 코스 추천 페이지 이미지가 새로고침해야만 뜨는 문제 수정
BodyPartLevelStep의 courseId prefetch가 coursesApi.getCourseDetail을
직접 불러 원본(raw) 응답을 캐시에 심어뒀는데, useCourseDetail은 같은
쿼리키로 매핑된 Course(이미지 asset key를 실제 경로로 해석한 값)를
캐싱한다 — 모양이 달라서 코스 추천 화면 첫 렌더가 미해석 원본을 잠깐
그리다 배경 리페치가 끝나야 정상으로 바뀌었다(체감상 "새로고침해야
뜸").

Course/CourseStep/CourseStepExercise 타입과 getCourseDetail/
useCourseDetail을 entities/course로 옮겨 프리패치와 실제 조회가
정확히 같은 매핑 함수를 쓰도록 통일한다. features/screening-flow가
features/course-recommendation을 직접 참조하지 않고도(레포 규칙)
같은 캐시 모양을 보장할 수 있는 유일한 방법이라 이 쪽으로 정리했다.
