+++ src/features/course-recommendation/index.ts
@@ -1,3 +1,3 @@
 export { CourseRecommendation } from "./ui/CourseRecommendation";
-export type { Course, CourseStep, CourseStepExercise } from "./model/types";
-export { useCourseDetail } from "./model/use-course-detail";
+export type { Course, CourseStep, CourseStepExercise } from "@/entities/course";
+export { useCourseDetail } from "@/entities/course";
